requirements.txt
torch>=2.1.0
torchvision>=0.16.0
timm>=1.0.9
scikit-learn>=1.3.0
pillow>=10.0.0
opencv-python>=4.8.0.76
matplotlib>=3.8.0
PySide6>=6.6.0
tqdm>=4.66.0
numpy==1.26.4
